refactor calc working times per ids
